Move checkpoint files
