+++ sui_core/src/lib.rs
@@ -16,3 +16,4 @@ pub mod make;
 pub mod safe_client;
 pub mod sui_json;
 pub mod transaction_input_checker;
+pub mod checkpoints;
